Safari-609.1.11.1/LayoutTests/media/W3C/audio/preload/preload_reflects_bogus_value-expected.txt

audio.preload - reflection test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "auto" is "auto"

TEST COMPLETE
spec reference
